requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydub>=0.25.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
